crosshap: Local Haplotype Clustering and Visualization

A local haplotyping visualization toolbox to capture major patterns of co-inheritance between clusters of linked variants, whilst connecting findings to phenotypic and demographic traits across individuals. 'crosshap' enables users to explore and understand genomic variation across a trait-associated region. For an example of successful local haplotype analysis, see Marsh et al. (2022) doi:10.1007/s00122-022-04045-8 .

See also

Author

Maintainer: Jacob Marsh jake.marsh@live.com.au (ORCID)

Authors:

Brady Johnston brady.johnston@me.com (ORCID)

Jakob Petereit petinho86@gmail.com (ORCID)