Visualize haplotypes

crosshap_viz() builds five individual plots using various elements of a HapObject created by run_haplotyping(). The central dotplot displays relationship between clusters of linked SNPs (marker groups), and distinct haplotypes present within the population. Vertical plots (top/bottom) visualize individuals and populations, grouped by haplotype. Horizontal plots (left/right) visualize SNP information, grouped by marker group cluster.

crosshap_viz(
  HapObject,
  plot_left = "allele",
  plot_right = "cluster",
  hide_labels = F
)

Arguments

HapObject

Haplotype object created by run_haplotyping().

plot_left

When plot_left = "allele", SNP allele frequency information is displayed, when plot_left = "pos", SNP position information is displayed.

plot_right

When plot_right = "pheno", phenotype associations for SNPs are displayed, when plot_right = "cluster", internal marker group linkage is displayed.

hide_labels

When TRUE, legends from plots are hidden.

Value

A patchwork object.