Cluster SNPs and identify haplotypes with HDBscan

Performs density-based clustering of SNPs in region of interest to identify marker groups. Defines haplotype combinations and classifies individuals based on characteristic combinations of marker group alleles. Returns a comprehensive haplotyping object, required to build clustering tree and visualize haplotypes.

run_hdbscan_haplotyping(
  vcf,
  LD,
  pheno,
  MGmin,
  minHap = 5,
  hetmiss_as = "allele",
  metadata = NULL,
  keep_outliers = F
)

Arguments

vcf

Input VCF for region of interest.

LD

Pairwise correlation matrix of SNPs in region from PLINK.

pheno

Input numeric phenotype data for each individual.

MGmin

Minimum SNPs in marker groups, MinPts parameter for HDBscan.

minHap

Minimum nIndividuals to keep haplotype combinations.

hetmiss_as

Treat heterozygous missing './X' as missing or allele.

metadata

Metadata input

keep_outliers

Smoothing